requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
pkg-resources==0.0.0
python-magic==0.4.15
requests==2.18.4
urllib3==1.22
